+++ dask_geomodeling/tests/test_raster.py
@@ -1663,6 +1663,24 @@ class TestBase(unittest.TestCase):
         self.assertEqual(view.get_data(**self.meta_request)["meta"], self.expected_meta)
         self.assertEqual(view.get_data(**self.time_request)["time"], self.expected_time)
 
+        # the 'value' determines the dtype. 1000 becomes uint16.
+        view = raster.Mask(store=self.raster, value=1000)
+        data = view.get_data(**self.vals_request)
+        self.assertEqual(str(view.dtype), "uint16")
+        assert_equal(data["values"], 1000)
+
+        # -1000 becomes int16.
+        view = raster.Mask(store=self.raster, value=-1000)
+        data = view.get_data(**self.vals_request)
+        self.assertEqual(str(view.dtype), "int16")
+        assert_equal(data["values"], -1000)
+
+        # 3.14159 becomes float32.
+        view = raster.Mask(store=self.raster, value=3.14159)
+        data = view.get_data(**self.vals_request)
+        self.assertEqual(str(view.dtype), "float32")
+        assert_equal(data["values"], 3.14159)
+
     def test_mask_below(self):
         # filled result
         view = raster.MaskBelow(store=self.raster, value=0)